Move AGENTS.md module into agents-md/ folder
- Group the module under its own folder (agent/agents-md/index.ts) matching the layout convention used by stream-handlers/, skill-tool/, etc.
- Update three call sites (AgentExecutor, activator, rpc-manager) and the module's internal relative paths to match the new depth.
